# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Japanese Vocabulary and Kanji Quiz API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}