# lib/api-zod/src/generated/types/vocabInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Japanese Vocabulary and Kanji Quiz API
 * OpenAPI spec version: 0.1.0
 */
import type { ExampleSentence } from './exampleSentence';
import type { GrammarExample } from './grammarExample';
import type { VocabInputLevel } from './vocabInputLevel';

export interface VocabInput {
  word: string;
  reading: string;
  romaji?: string;
  meaning: string;
  type: string;
  level: VocabInputLevel;
  section: number;
  chapter: number;
  example_sentences?: ExampleSentence[];
  grammar_examples?: GrammarExample[];
}